requirements.txt
numpy>=1.26.4
matplotlib>=3.9.2
scipy>=1.13.1
pandas>=2.2.3
jpype1>=1.5.2
tqdm>=4.67.1
mpi4py>=3.1.4
setuptools>=72.1.0